Fix license header
When the file contains no using directive, the existing method of adding
the license header would add it as a trailing trivia of the first member
(the namespace declaration).

I.e. It would appear at the end of the file.

Unfortunately, this is totally untestable, because it requires
osu!framework namespacing...